move proxyAdmin to common
