+++ packages/update-monitor/src/projects/synapse/index.ts
@@ -1,14 +1,14 @@
 import { providers } from 'ethers'
 
 import { getGnosisSafe } from '../../common/gnosisSafe'
+import { getProxyAdmin } from '../../common/proxyAdmin'
 import { DiscoveryEngine } from '../../discovery/DiscoveryEngine'
 import { ProjectParameters } from '../../types'
 import { addresses } from './constants'
-import { getProxyAdmin } from './contracts/proxyAdmin'
 import { getSynapseBridge } from './contracts/synapseBridge'
 import { getTimelock } from './contracts/timelock'
 
-export const SYNAPSE_NAME = 'synapse'
+export const SYNAPSE_NAME = 'synapseBridge'
 
 export async function getSynapseParameters(
   provider: providers.JsonRpcProvider,
@@ -18,7 +18,7 @@ export async function getSynapseParameters(
     contracts: await Promise.all([
       getSynapseBridge(provider),
       getGnosisSafe(provider, addresses.multisig, 'MultiSig'),
-      getProxyAdmin(provider),
+      getProxyAdmin(provider, addresses.proxyAdmin, 'ProxyAdmin'),
       getTimelock(provider),
     ]),
   }
